split a few things up
